+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.1-alpha.7",
-  "contentSha256_12": "da1d1468a19d",
+  "pkgVersion": "1.1.1-alpha.8",
+  "contentSha256_12": "6e98f7eaa999",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.1-alpha.7.js"
+  "hacsFile": "nodalia-cards-1.1.1-alpha.8.js"
 };
-export const pkgVersion = "1.1.1-alpha.7";
-export const contentSha256_12 = "da1d1468a19d";
+export const pkgVersion = "1.1.1-alpha.8";
+export const contentSha256_12 = "6e98f7eaa999";
 export const file = "nodalia-cards.bundle.js";